+++ src/_variables.scss
@@ -12,3 +12,5 @@ $--footer-padding-y: 1rem;
 $--main-page-hieght: calc(100vh - #{$--navbar-height} - #{$--footer-heigth} + #{$--footer-padding-y});
 $--cool-gray: rgb(33,33,33);
 $--hover-color: #FAC000;
+$--font-family: 'Courier New', Courier, monospace;
+$--link-color: #4d76e8;
