+++ src/components/footer/styles.module.scss
@@ -10,7 +10,6 @@
     .footer_section_element, .footer_section_header, .footer_caption_element, .footer_caption_header, .footer_additional_text {
         color: $--font-color;
         text-decoration: none;
-        font-family: Verdana, Geneva, Tahoma, sans-serif;
         //do usuniecia pozniej
         -webkit-user-select: none;
         -ms-user-select: none;
