+++ src/pages/registration/Register/styles.module.scss
@@ -1,0 +1,41 @@
+// css do  rejestracji
+@import "../../../variables";
+
+.holder {
+    background-color: $--dark-background-color;
+    height: 100vh;
+    color: $--font-color;
+    font-family: $--font-family;
+
+    a{
+        color: $--link-color;
+        text-decoration: none;
+        //do usuniecia pozniej
+        -webkit-user-select: none;
+        -ms-user-select: none;
+        user-select: none;
+    }
+
+    .title {
+        font-size: 2rem;
+        font-weight: bold;
+    }
+
+    .signUp, .logIn {
+        display: flex;
+        text-align: center;
+        align-items: center;
+        flex-direction: column;
+
+        .input {
+            background-color: lightgray;
+            width: 2vw;
+            margin: 0.5rem;
+            text-align: center;
+            padding: 0.3rem;
+            border-radius: 1rem;
+            font-size: 1.5rem;
+            
+        }
+    }
+}
