--- src/pages/registration/styles.module.scss
@@ -1,22 +1,0 @@
-.title {
-  font-size: 2rem;
-  font-weight: bold;
-}
-.signUp, .logIn {
-  text-align: center;
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-}
-.input {
-  background-color: lightgray;
-  width: 20vw;
-  margin: 0.5rem auto;
-  text-align: center;
-  padding: 0.3rem;
-  border-radius: 1rem;
-  font-size: 1.5rem;
-}
-#submit {
-  background-color: green;
-}
